+++ types/ContextMenu.d.ts
@@ -3,6 +3,7 @@ import ContextMenuOption from './interface/ContextMenuOption';
 import MenuOption from './interface/MenuOption';
 export interface MenuWrapper {
     show(e: MouseEvent, payload: any): void;
+    destroy(): void;
 }
 export default class ContextMenu {
     storeMenus: Menu[];
@@ -15,4 +16,5 @@ export default class ContextMenu {
     create(option: MenuOption): MenuWrapper;
     showMenu(e: MouseEvent, menu: Menu, payload: any): void;
     hideAllMenu(): void;
+    destroy(menu: Menu): void;
 }