+++ types/Menu.d.ts
@@ -1,26 +1,24 @@
 import MenuItem from './MenuItem';
 import MenuItemOption from './interface/MenuItemOption';
 import MenuOption from './interface/MenuOption';
-export interface InnerOption {
-    position?: string;
-}
-export default class Menu {
+import Panel, { PanelOption } from './Panel';
+export default class Menu extends Panel {
+    ul: HTMLElement;
     level: number;
-    el: HTMLElement;
-    width: number;
-    height: number;
     items: MenuItemOption[];
     children: MenuItem[];
     payload: any;
-    innerOption: InnerOption;
-    constructor(level: number, option: MenuOption, innerOption?: InnerOption);
+    constructor(level: number, option: MenuOption, panelOption?: PanelOption);
     init(): void;
     addChildren(items: MenuItemOption[]): void;
-    show(e: MouseEvent, payload: any): void;
-    calcPosition(e: MouseEvent): void;
-    hide(): void;
+    show(e: MouseEvent, payload?: any): void;
+    calcPosition(e: MouseEvent): {
+        x: number;
+        y: number;
+    };
     removeAllHover(): void;
     removeChildMenus(): void;
     removeItemHover(): void;
     closeAllMenus(): void;
+    destroy(): void;
 }