+++ types/index.d.ts
@@ -1,1 +1,3 @@
 export { default } from './ContextMenu';
+export { default as Panel } from './Panel';
+export { default as h } from './utils/h';